Add files via upload
The organization operates on Pay as you go basis on higher purchase. Once a customer acquires a device in form of a loan, their loan repayment rate is tracked. This will enable the organization to follow up on the loan and inform on whether the customer will be eligible for another form of loan in the future.